Move D/I Mem tests to MemorySpec
