+++ src/test/scala/chiselTests/experimental/hierarchy/Annotations.scala
@@ -8,7 +8,7 @@ import chisel3.{Data, HasTarget, MemBase}
 import chisel3.experimental.hierarchy.{Definition, Hierarchy, Instance}
 
 // These annotations exist purely for testing purposes
-private[hierarchy] object Annotations {
+object Annotations {
   case class MarkAnnotation(target: IsMember, tag: String) extends SingleTargetAnnotation[IsMember] {
     def duplicate(n: IsMember): Annotation = this.copy(target = n)
   }
